requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
altair==5.1.1
    # via streamlit
anyio==4.0.0
    # via jupyter-server
argon2-cffi==23.1.0
    # via jupyter-server
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
arrow==1.2.3
    # via isoduration
asttokens==2.4.0
    # via stack-data
async-lru==2.0.4
    # via jupyterlab
attrs==23.1.0
    # via
    #   jsonschema
    #   referencing
babel==2.12.1
    # via jupyterlab-server
backcall==0.2.0
    # via ipython
beautifulsoup4==4.12.2
    # via nbconvert
black==23.7.0
    # via -r requirements.in
bleach==6.0.0
    # via nbconvert
blinker==1.6.2
    # via streamlit
cachetools==5.3.1
    # via streamlit
certifi==2023.7.22
    # via requests
cffi==1.15.1
    # via argon2-cffi-bindings
charset-normalizer==3.2.0
    # via requests
click==8.1.7
    # via
    #   black
    #   streamlit
comm==0.1.4
    # via
    #   ipykernel
    #   ipywidgets
debugpy==1.7.0
    # via ipykernel
decorator==5.1.1
    # via ipython
defusedxml==0.7.1
    # via nbconvert
executing==1.2.0
    # via stack-data
fastjsonschema==2.18.0
    # via nbformat
fqdn==1.5.1
    # via jsonschema
gitdb==4.0.10
    # via gitpython
gitpython==3.1.35
    # via streamlit
idna==3.4
    # via
    #   anyio
    #   jsonschema
    #   requests
importlib-metadata==6.8.0
    # via streamlit
iniconfig==2.0.0
    # via pytest
ipykernel==6.25.2
    # via jupyterlab
ipython==8.15.0
    # via
    #   -r requirements.in
    #   ipykernel
    #   ipywidgets
ipywidgets==8.1.0
    # via -r requirements.in
isoduration==20.11.0
    # via jsonschema
isort==5.12.0
    # via -r requirements.in
jedi==0.19.0
    # via ipython
jinja2==3.1.2
    # via
    #   altair
    #   jupyter-server
    #   jupyterlab
    #   jupyterlab-server
    #   nbconvert
    #   pydeck
json5==0.9.14
    # via jupyterlab-server
jsonpointer==2.4
    # via jsonschema
jsonschema[format-nongpl]==4.19.0
    # via
    #   altair
    #   jupyter-events
    #   jupyterlab-server
    #   nbformat
jsonschema-specifications==2023.7.1
    # via jsonschema
jupyter-client==8.3.1
    # via
    #   ipykernel
    #   jupyter-server
    #   nbclient
jupyter-core==5.3.1
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-server
    #   jupyterlab
    #   nbclient
    #   nbconvert
    #   nbformat
jupyter-events==0.7.0
    # via jupyter-server
jupyter-lsp==2.2.0
    # via jupyterlab
jupyter-server==2.7.3
    # via
    #   jupyter-lsp
    #   jupyterlab
    #   jupyterlab-server
    #   notebook-shim
jupyter-server-terminals==0.4.4
    # via jupyter-server
jupyterlab==4.0.5
    # via -r requirements.in
jupyterlab-pygments==0.2.2
    # via nbconvert
jupyterlab-server==2.24.0
    # via jupyterlab
jupyterlab-widgets==3.0.8
    # via ipywidgets
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.3
    # via
    #   jinja2
    #   nbconvert
matplotlib-inline==0.1.6
    # via
    #   ipykernel
    #   ipython
mdurl==0.1.2
    # via markdown-it-py
mistune==3.0.1
    # via nbconvert
mypy-extensions==1.0.0
    # via black
nbclient==0.8.0
    # via nbconvert
nbconvert==7.8.0
    # via jupyter-server
nbformat==5.9.2
    # via
    #   jupyter-server
    #   nbclient
    #   nbconvert
nest-asyncio==1.5.7
    # via ipykernel
notebook-shim==0.2.3
    # via jupyterlab
numpy==1.25.2
    # via
    #   altair
    #   pandas
    #   patsy
    #   plotly-express
    #   pyarrow
    #   pydeck
    #   scipy
    #   statsmodels
    #   streamlit
    #   streamlit-cropper
overrides==7.4.0
    # via jupyter-server
packaging==23.1
    # via
    #   altair
    #   black
    #   ipykernel
    #   jupyter-server
    #   jupyterlab
    #   jupyterlab-server
    #   nbconvert
    #   plotly
    #   pytesseract
    #   pytest
    #   statsmodels
    #   streamlit
pandas==2.1.0
    # via
    #   altair
    #   plotly-express
    #   statsmodels
    #   streamlit
pandocfilters==1.5.0
    # via nbconvert
parso==0.8.3
    # via jedi
pathspec==0.11.2
    # via black
patsy==0.5.3
    # via
    #   plotly-express
    #   statsmodels
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==9.5.0
    # via
    #   pytesseract
    #   streamlit
    #   streamlit-cropper
platformdirs==3.10.0
    # via
    #   black
    #   jupyter-core
plotly==5.16.1
    # via plotly-express
plotly-express==0.4.1
    # via -r requirements.in
pluggy==1.3.0
    # via pytest
prometheus-client==0.17.1
    # via jupyter-server
prompt-toolkit==3.0.39
    # via ipython
protobuf==4.24.3
    # via streamlit
psutil==5.9.5
    # via ipykernel
ptyprocess==0.7.0
    # via
    #   pexpect
    #   terminado
pure-eval==0.2.2
    # via stack-data
pyarrow==13.0.0
    # via streamlit
pycparser==2.21
    # via cffi
pydeck==0.8.0
    # via streamlit
pygments==2.16.1
    # via
    #   ipython
    #   nbconvert
    #   rich
pympler==1.0.1
    # via streamlit
pytesseract==0.3.10
    # via -r requirements.in
pytest==7.4.2
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   arrow
    #   jupyter-client
    #   pandas
    #   streamlit
python-json-logger==2.0.7
    # via jupyter-events
pytz==2023.3.post1
    # via pandas
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
pyyaml==6.0.1
    # via jupyter-events
pyzmq==25.1.1
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-server
referencing==0.30.2
    # via
    #   jsonschema
    #   jsonschema-specifications
    #   jupyter-events
requests==2.31.0
    # via
    #   -r requirements.in
    #   jupyterlab-server
    #   streamlit
rfc3339-validator==0.1.4
    # via
    #   jsonschema
    #   jupyter-events
rfc3986-validator==0.1.1
    # via
    #   jsonschema
    #   jupyter-events
rich==13.5.2
    # via streamlit
rpds-py==0.10.2
    # via
    #   jsonschema
    #   referencing
scipy==1.11.2
    # via
    #   plotly-express
    #   statsmodels
send2trash==1.8.2
    # via jupyter-server
six==1.16.0
    # via
    #   asttokens
    #   bleach
    #   patsy
    #   python-dateutil
    #   rfc3339-validator
smmap==5.0.0
    # via gitdb
sniffio==1.3.0
    # via anyio
soupsieve==2.5
    # via beautifulsoup4
stack-data==0.6.2
    # via ipython
statsmodels==0.14.0
    # via plotly-express
streamlit==1.26.0
    # via
    #   -r requirements.in
    #   streamlit-cropper
streamlit-cropper==0.2.1
    # via -r requirements.in
tenacity==8.2.3
    # via
    #   plotly
    #   streamlit
terminado==0.17.1
    # via
    #   jupyter-server
    #   jupyter-server-terminals
tinycss2==1.2.1
    # via nbconvert
toml==0.10.2
    # via streamlit
toolz==0.12.0
    # via altair
tornado==6.3.3
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-server
    #   jupyterlab
    #   streamlit
    #   terminado
traitlets==5.9.0
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-core
    #   jupyter-events
    #   jupyter-server
    #   jupyterlab
    #   matplotlib-inline
    #   nbclient
    #   nbconvert
    #   nbformat
typing-extensions==4.7.1
    # via streamlit
tzdata==2023.3
    # via
    #   pandas
    #   pytz-deprecation-shim
tzlocal==4.3.1
    # via streamlit
uri-template==1.3.0
    # via jsonschema
urllib3==2.0.4
    # via requests
validators==0.22.0
    # via streamlit
watchdog==3.0.0
    # via streamlit
wcwidth==0.2.6
    # via prompt-toolkit
webcolors==1.13
    # via jsonschema
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
websocket-client==1.6.2
    # via jupyter-server
widgetsnbextension==4.0.8
    # via ipywidgets
zipp==3.16.2
    # via importlib-metadata
